+++ mne/gui/_viewer.py
@@ -14,12 +14,14 @@ from mayavi.tools.mlab_scene_model import MlabSceneModel
 from pyface.api import error
 from traits.api import (HasTraits, HasPrivateTraits, on_trait_change,
                         Instance, Array, Bool, Button, Enum, Float, Int, List,
-                        Range, Str, RGBColor)
+                        Range, Str, RGBColor, Property, cached_property)
 from traitsui.api import View, Item, HGroup, VGrid, VGroup
 from tvtk.api import tvtk
 
 from ..defaults import DEFAULTS
-from ..surface import complete_surface_info, _project_onto_surface
+from ..surface import (complete_surface_info, _project_onto_surface,
+                       _normalize_vectors)
+from ..source_space import _points_outside_surface
 from ..transforms import apply_trans
 from ..utils import SilenceStdout
 from ..viz._3d import _create_mesh_surf, _toggle_mlab_render
@@ -128,19 +130,27 @@ class Object(HasPrivateTraits):
     # projection onto a surface
     project_to_points = Array(float, shape=(None, 3))
     project_to_tris = Array(int, shape=(None, 3))
-    project_to_surface = Bool(False, label='Project')
+    project_to_surface = Bool(False, label='Project', desc='project points '
+                              'onto the surface')
+    orient_to_surface = Bool(False, label='Orient', desc='orient points '
+                             'toward the surface')
+    scale_by_distance = Bool(False, label='Dist.', desc='scale points by '
+                             'distance from the surface')
+    mark_inside = Bool(False, label='Mark', desc='mark points inside the '
+                       'surface in a different color')
 
     scene = Instance(MlabSceneModel, ())
     src = Instance(VTKDataSource)
 
     # This should be Tuple, but it is broken on Anaconda as of 2016/12/16
     color = RGBColor((1., 1., 1.))
+    inside_color = RGBColor((0., 0., 0.))
     point_scale = Float(10, label='Point Scale')
     opacity = Range(low=0., high=1., value=1.)
     visible = Bool(True)
 
     # don't put project_to_tris here, just always set project_to_points second
-    @on_trait_change('trans,points,project_to_surface,project_to_points')
+    @on_trait_change('trans,points,project_to_surface,mark_inside')
     def _update_points(self):
         """Update the location of the plotted points."""
         if not hasattr(self.src, 'data'):
@@ -164,23 +174,40 @@ class Object(HasPrivateTraits):
             pts = self.points
 
         # Do the projection if required
-        if self.project_to_surface and len(self.project_to_points) > 1:
+        if len(self.project_to_points) > 1 and len(pts) > 0:
             surf = dict(rr=np.array(self.project_to_points),
                         tris=np.array(self.project_to_tris))
             method = 'accurate' if len(surf['rr']) <= 20484 else 'nearest'
-            pts, nn = _project_onto_surface(
+            proj_pts, proj_nn = _project_onto_surface(
                 pts, surf, project_rrs=True, return_nn=True,
                 method=method)[2:4]
-            self.src.data.point_data.normals = nn
-        self.src.data.points = pts
+            vec = pts - proj_pts  # point to the surface
+            if self.project_to_surface:
+                pts = proj_pts
+                nn = proj_nn
+            else:
+                nn = vec.copy()
+                _normalize_vectors(nn)
+            if self.mark_inside and not self.project_to_surface:
+                scalars = _points_outside_surface(pts, surf).astype(int)
+            else:
+                scalars = np.ones(len(pts))
+            # With this, a point exactly on the surface is of size point_scale
+            dist = np.linalg.norm(vec, axis=-1, keepdims=True)
+            self.src.data.point_data.normals = (250 * dist + 1) * nn
+            self.src.data.point_data.scalars = scalars
+            self.glyph.actor.mapper.scalar_range = [0., 1.]
         self.src.data.point_data.update()
+        self.src.data.points = pts
         return True
 
 
 class PointObject(Object):
     """Represent a group of individual points in a mayavi scene."""
 
-    label = Bool(False, enabled_when='visible')
+    label = Bool(False)
+    projectable = Bool(False)  # set based on type of points
+    orientable = Property(depends_on=['project_to_points', 'project_to_tris'])
     text3d = List
 
     glyph = Instance(Glyph)
@@ -190,7 +217,7 @@ class PointObject(Object):
                        Item('color', show_label=False),
                        Item('opacity')))
 
-    def __init__(self, view='points', projectable=False, *args, **kwargs):
+    def __init__(self, view='points', *args, **kwargs):
         """Init.
 
         Parameters
@@ -200,12 +227,17 @@ class PointObject(Object):
             or a point cloud.
         """
         self._view = view
-        self._projectable = projectable
         super(PointObject, self).__init__(*args, **kwargs)
 
     def default_traits_view(self):  # noqa: D102
         color = Item('color', show_label=False)
         scale = Item('point_scale', label='Size')
+        orient = Item('orient_to_surface',
+                      enabled_when='orientable and not project_to_surface')
+        dist = Item('scale_by_distance',
+                    enabled_when='orientable and not project_to_surface')
+        mark = Item('mark_inside',
+                    enabled_when='orientable and not project_to_surface')
         if self._view == 'points':
             visible = Item('visible', label='Show', show_label=True)
             views = (visible, color, scale, 'label')
@@ -214,9 +246,10 @@ class PointObject(Object):
             views = (visible, color, scale)
         else:
             raise ValueError("PointObject(view = %r)" % self._view)
-        if self._projectable:
-            views = views + (Item('project_to_surface', show_label=True),)
-        return View(HGroup(*views))
+        group2 = HGroup(dist, Item('project_to_surface', show_label=True,
+                                   enabled_when='projectable'),
+                        orient, mark, show_left=False)
+        return View(HGroup(HGroup(*views), group2))
 
     @on_trait_change('label')
     def _show_labels(self, show):
@@ -263,6 +296,12 @@ class PointObject(Object):
                                mode='sphere')
         glyph.actor.property.backface_culling = True
         glyph.glyph.glyph.vector_mode = 'use_normal'
+        glyph.glyph.glyph.clamping = False
+
+        glyph.actor.mapper.color_mode = 'map_scalars'
+        glyph.actor.mapper.scalar_mode = 'use_point_data'
+        glyph.actor.mapper.use_lookup_table_scalar_range = False
+
         self.src = scatter
         self.glyph = glyph
 
@@ -270,21 +309,39 @@ class PointObject(Object):
         self.sync_trait('color', self.glyph.actor.property, mutual=False)
         self.sync_trait('visible', self.glyph)
         self.sync_trait('opacity', self.glyph.actor.property)
+        self.sync_trait('mark_inside', self.glyph.actor.mapper,
+                        'scalar_visibility')
         self.on_trait_change(self._update_points, 'points')
-        self._update_project_to_surface()
+        self._update_markers()
+        self._update_colors()
         _toggle_mlab_render(self, True)
+        # self.scene.camera.parallel_scale = _scale
 
-#         self.scene.camera.parallel_scale = _scale
+    @on_trait_change('color,inside_color')
+    def _update_colors(self):
+        if self.glyph is None:
+            return
+        # inside_color is the surface color, let's try to get far
+        # from that
+        inside = np.array(self.inside_color)
+        # if it's too close to gray, just use black:
+        if np.mean(np.abs(inside - 0.5)) < 0.2:
+            inside.fill(0.)
+        else:
+            inside = 1 - inside
+        colors = np.array([tuple(inside) + (1,),
+                           tuple(self.color) + (1,)]) * 255.
+        self.glyph.module_manager.scalar_lut_manager.lut.table = colors
 
-    @on_trait_change('project_to_surface')
-    def _update_project_to_surface(self):
-        if not self.glyph:
+    @on_trait_change('project_to_surface,orient_to_surface,scale_by_distance')
+    def _update_markers(self):
+        if self.glyph is None:
             return
         defaults = DEFAULTS['coreg']
         gs = self.glyph.glyph.glyph_source
         res = getattr(gs.glyph_source, 'theta_resolution',
                       getattr(gs.glyph_source, 'resolution', None))
-        if self.project_to_surface:
+        if self.project_to_surface or self.orient_to_surface:
             gs.glyph_source = tvtk.CylinderSource()
             gs.glyph_source.height = defaults['eegp_height']
             gs.glyph_source.center = (0., -defaults['eegp_height'], 0)
@@ -293,7 +350,10 @@ class PointObject(Object):
             gs.glyph_source = tvtk.SphereSource()
             gs.glyph_source.phi_resolution = res
             gs.glyph_source.theta_resolution = res
-        self.glyph.glyph.scale_mode = 'data_scaling_off'
+        if self.scale_by_distance:
+            self.glyph.glyph.scale_mode = 'scale_by_vector'
+        else:
+            self.glyph.glyph.scale_mode = 'data_scaling_off'
 
     def _resolution_changed(self, new):
         if not self.glyph:
@@ -305,6 +365,15 @@ class PointObject(Object):
             gs.theta_resolution = new
         else:
             gs.resolution = new
+
+    @cached_property
+    def _get_orientable(self):
+        orientable = (len(self.project_to_points) > 0 and
+                      len(self.project_to_tris) > 0)
+        return orientable
+
+
+# XXX eventually we should update the normals whenever "points" is changed
 
 
 class SurfaceObject(Object):