+++ doc/whats_new.rst
@@ -77,7 +77,7 @@ Changelog
 
 - Add ability to exclude components interactively by clicking on their labels in :meth:`mne.preprocessing.ICA.plot_components` by `Mikołaj Magnuski`_
 
-- Add projection of EEG electrodes to the head surface in ``mne coreg`` / :func:`mne.gui.coregistration` by `Eric Larson`_
+- Add projection of EEG electrodes, orientation of extra points, and scaling points by distance to the head surface; display of HPI points; and marking points inside the head surface in a different color by ``mne coreg`` / :func:`mne.gui.coregistration` by `Eric Larson`_
 
 - Add reader for manual annotations of raw data produced by Brainstorm by `Anne-Sophie Dubarry`_